added simplified comparison example
Annelies told me the html was getting complicated (lots of code to highlight a word). This example demonstrates a method which only needs adding <span class="hl#">some text</span> around the highlighted text pairs, replacing # with the same number per pair (number of text, but start with hl...)(it is HL for highlight, not h1). Javascript adds and removes the .highlight css class on mouse events. css and javascript code can be in seperate files of course.